elimina endpoint obsoleto e migra outro"
